feat(top3): Add top3 feature
