+++ src/features/leaderboard/top3/index.ts
@@ -1,0 +1,1 @@
+export { default as TopThreeReveal } from './components/TopThreeBoard';
