+++ src/features/leaderboard/top3/types/CardStyles.ts
@@ -1,0 +1,6 @@
+export interface CardStyles {
+  border: string;
+  bg: string;
+  text: string;
+  badge: string;
+}
