+++ src/features/leaderboard/top3/types/Contestant.ts
@@ -1,0 +1,5 @@
+export interface Contestant {
+  id: number;
+  fullname: string;
+  score: number;
+}
