+++ src/features/leaderboard/top3/utils/rippleUtils.ts
@@ -1,0 +1,9 @@
+export const createRipple = (color: string) => {
+  const rippleContainer = document.createElement('div');
+  rippleContainer.className = `ripple-container ${color}-ripple`;
+  document.body.appendChild(rippleContainer);
+
+  setTimeout(() => {
+    document.body.removeChild(rippleContainer);
+  }, 2000);
+};
