+++ src/features/leaderboard/top3/utils/soundUtils.ts
@@ -1,0 +1,6 @@
+export const playSound = (type: string, isMuted: boolean) => {
+  if (!isMuted) {
+    const audio = new Audio(`/audio/${type}.mp3`);
+    audio.play().catch((err) => console.error('Audio error:', err));
+  }
+};
